refactors bot typescript
