+++ interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './discord.interface'
+export * from './supabase.interface'
+export * from './xbox.interface'
