+++ interfaces/xbox.interface.ts
@@ -1,4 +1,3 @@
-
 interface Image {
   URI: string,
   Width: number,
@@ -10,21 +9,6 @@ interface Price {
   SalesPrice: string,
   IsFree: boolean
 } 
-
-export interface Subscription {
-  channel: string;
-  webhook: string;
-}
-
-export interface Game {
-  id: string,
-  title: string,
-  developer: string,
-  image: string | null,
-  price: number | null,
-  size: number | null,
-  description: string
-}
 
 export interface XboxGame {
   ProductTitle: string,
